Clean up of warzone two generator
